[UDT-72] refactor: 카드 컴포넌트 변경 및 폴더 변경
